Make ScheduledSubscription public, expose scheduler field
